mslearn-ai-lenguage/07-speech/requirements.txt
azure-cognitiveservices-speech==1.30.0
playsound==1.2.2
python-dotenv==1.0.1
streamlit==1.42.0